requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.14.0
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
whitenoise==6.6.0
